move to SnippetUtils + change comma addition and sort logic
